requirements.txt
Django==5.2.8
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0

